refactor(admin-cb): move block wrappers to components folder
