+++ src/admin/content-block/components/wrappers/index.ts
@@ -1,0 +1,4 @@
+export { default as MediaGridWrapper } from './MediaGridWrapper/MediaGridWrapper';
+export { default as MediaPlayerTitleTextButtonWrapper } from './MediaPlayerTitleTextButtonWrapper/MediaPlayerTitleTextButtonWrapper';
+export * from './MediaPlayerWrapper/MediaPlayerWrapper';
+export { default as PageOverviewWrapper } from './PageOverviewWrapper/PageOverviewWrapper';
