--- src/admin/content-block/helpers/wrappers/index.ts
@@ -1,3 +1,0 @@
-export { default as BlockMediaGridWrapper } from './BlockMediaGridWrapper';
-export * from './BlockMediaPlayerWrapper';
-export * from './BlockMediaPlayerTitleTextButtonWrapper';
